llm_service/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
google-generativeai
python-multipart
